Split consignments and contamination modules
Too much code accumulated in both, contamination is large and may be even larger.
